Adjust create mirror deb to work with Ubuntu 16.04.

+++ mirror/create_mirror_deb.sh
@@ -4,25 +4,36 @@ export DISTRO=$(cat /etc/*-release|grep ^ID\=|awk -F\= {'print $2'}|sed s/\"//g)
 [[ -z ${MIRROR_BUILD_DIR} ]] && export MIRROR_BUILD_DIR=${PWD}
 [[ -z ${MIRROR_OUTPUT_DIR} ]] && export MIRROR_OUTPUT_DIR=${PWD}/mirror-dist
 
-DEB_PACKAGE_LIST=$(<${MIRROR_BUILD_DIR}/dependencies/pnda-deb-package-dependencies.txt)
+OS_RELEASE=$(lsb_release -sc)
+OS_VERSION=$(lsb_release -sr)
+OS_ID=$(lsb_release -si)
+OS_SHORT=${OS_ID,,}${OS_VERSION%.*}
+
+DEB_PACKAGE_LIST=$(<${MIRROR_BUILD_DIR}/dependencies/pnda-deb-package-dependencies.txt.${OS_RELEASE})
+
+if [ ! -f "$DEB_PACKAGE_LIST" ];
+then
+    echo "File $DEB_PACKAGE_LIST does not exist."
+    exit 1
+fi
 
 export DEBIAN_FRONTEND=noninteractive
 DEB_REPO_DIR=$MIRROR_OUTPUT_DIR/mirror_deb
 
-echo 'deb [arch=amd64] https://archive.cloudera.com/cm5/ubuntu/trusty/amd64/cm/ trusty-cm5.12.1 contrib' > /etc/apt/sources.list.d/cloudera-manager.list
-curl -L 'https://archive.cloudera.com/cm5/ubuntu/trusty/amd64/cm/archive.key' | apt-key add -
+echo "deb [arch=amd64] https://archive.cloudera.com/cm5/ubuntu/${OS_RELEASE}/amd64/cm/ ${OS_RELEASE}-cm5.12.1 contrib" > /etc/apt/sources.list.d/cloudera-manager.list
+curl -L "https://archive.cloudera.com/cm5/ubuntu/${OS_RELEASE}/amd64/cm/archive.key" | apt-key add -
 
-echo 'deb [arch=amd64] http://repo.saltstack.com/apt/ubuntu/14.04/amd64/archive/2015.8.11/ trusty main' > /etc/apt/sources.list.d/saltstack.list
-curl -L 'http://repo.saltstack.com/apt/ubuntu/14.04/amd64/archive/2015.8.11/SALTSTACK-GPG-KEY.pub' | apt-key add -
+echo "deb [arch=amd64] http://repo.saltstack.com/apt/ubuntu/${OS_VERSION}/amd64/archive/2015.8.11/ ${OS_RELEASE} main" > /etc/apt/sources.list.d/saltstack.list
+curl -L "http://repo.saltstack.com/apt/ubuntu/${OS_VERSION}/amd64/archive/2015.8.11/SALTSTACK-GPG-KEY.pub" | apt-key add -
 
-echo 'deb http://public-repo-1.hortonworks.com/ambari/ubuntu14/2.x/updates/2.6.1.0 Ambari main' > /etc/apt/sources.list.d/ambari.list
+echo "deb http://public-repo-1.hortonworks.com/ambari/${OS_SHORT}/2.x/updates/2.6.1.0 Ambari main" > /etc/apt/sources.list.d/ambari.list
 apt-key adv --recv-keys --keyserver keyserver.ubuntu.com B9733A7A07513CAD
 
 apt-get -y update
 apt-get -y install apt-transport-https curl dpkg-dev debfoster rng-tools
 
-debconf-set-selections <<< 'mysql-server mysql-server/root_password password your_password'
-debconf-set-selections <<< 'mysql-server mysql-server/root_password_again password your_password'
+debconf-set-selections <<< 'mysql-server mysql-server/root_password password cisco'
+debconf-set-selections <<< 'mysql-server mysql-server/root_password_again password cisco'
 apt-get -y install $DEB_PACKAGE_LIST
 
 rm -rf $DEB_REPO_DIR
@@ -31,8 +42,8 @@ cd $DEB_REPO_DIR
 apt-get download $DEB_PACKAGE_LIST
 for DEB_PACKAGE in $DEB_PACKAGE_LIST
 do
-	DEBP=(${DEB_PACKAGE//=/ })
-	echo "managing sub deps for $DEBP"
+    DEBP=(${DEB_PACKAGE//=/ })
+    echo "managing sub deps for $DEBP"
     SUB_DEP_LIST=$(debfoster -d $DEBP | grep '^ ' || true)
     if ! [ -z "$SUB_DEP_LIST" ]; then apt-get download $SUB_DEP_LIST; fi
 done
